+++ ChatUI/.idea/artifacts/JavaFXApp.xml
@@ -8,7 +8,7 @@
     </properties>
     <root id="root">
       <element id="archive" name="JavaFXApp.jar">
-        <element id="module-output" name="ChatApp" />
+        <element id="module-output" name="ChatUI" />
       </element>
     </root>
   </artifact>
