+++ ChatUI/.idea/modules.xml
@@ -2,7 +2,7 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/ChatApp.iml" filepath="$PROJECT_DIR$/ChatApp.iml" />
+      <module fileurl="file://$PROJECT_DIR$/ChatUI.iml" filepath="$PROJECT_DIR$/ChatUI.iml" />
     </modules>
   </component>
 </project>
