+++ ChatUI/ChatUI.iml
@@ -8,6 +8,6 @@
     <orderEntry type="inheritedJdk" />
     <orderEntry type="sourceFolder" forTests="false" />
     <orderEntry type="library" name="JavaFX-11" level="application" />
-    <orderEntry type="library" name="lib" level="project" />
+    <orderEntry type="library" name="Jfoenix" level="application" />
   </component>
 </module>
