+++ ChatUI/src/Application/Main.java
@@ -1,0 +1,24 @@
+package Application;
+
+import javafx.application.Application;
+import javafx.fxml.FXMLLoader;
+import javafx.stage.Stage;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
+
+public class Main extends Application {
+
+    @Override
+    public void start(Stage primaryStage) throws Exception{
+        Parent root = FXMLLoader.load(getClass().getResource("/FXML/LoginUI.fxml"));
+        primaryStage.setTitle("Seen");
+        primaryStage.setScene(new Scene(root, 600, 400));
+        primaryStage.show();
+        primaryStage.setResizable(false);
+    }
+
+
+    public static void main(String[] args) {
+        launch(args);
+    }
+}
