+++ ChatUI/src/Controller/LoginController.java
@@ -1,4 +1,16 @@
 package Controller;
 
-public class LoginController {
+import java.net.URL;
+import java.util.ResourceBundle;
+
+import javafx.fxml.FXML;
+import javafx.fxml.Initializable;
+
+public class LoginController implements Initializable{
+
+
+    @Override
+    public  void initialize(URL arg0, ResourceBundle arg1) {
+       //
+    }
 }
