--- ChatUI/src/FXML/Login.fxml
@@ -1,64 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.CheckBox?>
-<?import javafx.scene.control.PasswordField?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.image.Image?>
-<?import javafx.scene.image.ImageView?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.text.Font?>
-
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" style="-fx-background-color: #121212;" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1">
-   <children>
-      <TextField layoutX="160.0" layoutY="125.0" prefHeight="28.0" prefWidth="280.0" promptText="Username" style="-fx-border-color: #121212 #121212 #FFFFFF #121212; -fx-background-color: #121212;">
-         <font>
-            <Font size="14.0" />
-         </font>
-      </TextField>
-      <PasswordField layoutX="160.0" layoutY="185.0" prefHeight="28.0" prefWidth="280.0" promptText="Password" style="-fx-border-color: #121212 #121212 #FFFFFF #121212; -fx-background-color: #121212;">
-         <font>
-            <Font size="14.0" />
-         </font>
-      </PasswordField>
-      <Button layoutX="486.0" layoutY="14.0" mnemonicParsing="false" prefHeight="40.0" prefWidth="100.0" style="-fx-background-color: #03DAC6;" text="Sign Up">
-         <font>
-            <Font size="14.0" />
-         </font>
-      </Button>
-      <CheckBox layoutX="163.0" layoutY="243.0" mnemonicParsing="false" text="Remember me" textFill="WHITE">
-         <font>
-            <Font size="15.0" />
-         </font>
-      </CheckBox>
-      <Button layoutX="340.0" layoutY="234.0" mnemonicParsing="false" prefHeight="40.0" prefWidth="100.0" style="-fx-background-color: #03DAC6;" text="Login">
-         <font>
-            <Font size="14.0" />
-         </font>
-      </Button>
-      <Button layoutX="202.0" layoutY="284.0" mnemonicParsing="false" prefHeight="40.0" prefWidth="197.0" style="-fx-background-color: #121212;" text="Forgot Your Password ?" textFill="WHITE" underline="true">
-         <font>
-            <Font size="14.0" />
-         </font>
-      </Button>
-      <ImageView fitHeight="95.0" fitWidth="100.0" layoutX="253.0" layoutY="7.0" pickOnBounds="true" preserveRatio="true">
-         <image>
-            <Image url="@../../assets/tick.png" />
-         </image></ImageView>
-      <ImageView fitHeight="48.0" fitWidth="50.0" layoutX="92.0" layoutY="116.0" pickOnBounds="true" preserveRatio="true">
-         <image>
-            <Image url="@../../assets/user.png" />
-         </image>
-      </ImageView>
-      <ImageView fitHeight="48.0" fitWidth="48.0" layoutX="92.0" layoutY="176.0" pickOnBounds="true" preserveRatio="true">
-         <image>
-            <Image url="@../../assets/padlock.png" />
-         </image>
-      </ImageView>
-      <ImageView fitHeight="68.0" fitWidth="78.0" layoutX="266.0" layoutY="318.0" pickOnBounds="true" preserveRatio="true">
-         <image>
-            <Image url="@../../assets/Ellipsis-2s-200px.gif" />
-         </image>
-      </ImageView>
-   </children>
-</AnchorPane>
