+++ ChatUI/src/FXML/RegisterUI.fxml
@@ -1,0 +1,61 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import com.jfoenix.controls.JFXButton?>
+<?import com.jfoenix.controls.JFXPasswordField?>
+<?import com.jfoenix.controls.JFXTextField?>
+<?import javafx.scene.image.Image?>
+<?import javafx.scene.image.ImageView?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.text.Font?>
+
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="450.0" prefWidth="600.0" style="-fx-background-color: #121212;" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1">
+    <children>
+        <ImageView fitHeight="95.0" fitWidth="100.0" layoutX="253.0" layoutY="14.0" pickOnBounds="true" preserveRatio="true">
+            <image>
+                <Image url="@../assets/tick.png" />
+            </image></ImageView>
+        <JFXTextField focusColor="#03dac6" layoutX="152.0" layoutY="125.0" prefWidth="300.0" promptText="Username" style="-fx-text-fill: #ffffff;">
+            <font>
+                <Font size="16.0" />
+            </font></JFXTextField>
+        <JFXPasswordField focusColor="#03dac6" layoutX="152.0" layoutY="181.0" prefWidth="300.0" promptText="Password" style="-fx-text-fill: #ffffff;">
+            <font>
+                <Font size="16.0" />
+            </font></JFXPasswordField>
+        <ImageView fitHeight="95.0" fitWidth="100.0" layoutX="253.0" layoutY="341.0" pickOnBounds="true" preserveRatio="true">
+            <image>
+                <Image url="@../assets/loading.gif" />
+            </image>
+        </ImageView>
+        <ImageView fitHeight="42.0" fitWidth="50.0" layoutX="108.0" layoutY="124.0" pickOnBounds="true" preserveRatio="true">
+            <image>
+                <Image url="@../assets/user.png" />
+            </image>
+        </ImageView>
+        <ImageView fitHeight="42.0" fitWidth="50.0" layoutX="108.0" layoutY="179.0" pickOnBounds="true" preserveRatio="true">
+            <image>
+                <Image url="@../assets/lock.png" />
+            </image>
+        </ImageView>
+        <JFXButton layoutX="201.0" layoutY="301.0" prefHeight="40.0" prefWidth="200.0" style="-fx-background-color: #03dac6;" text="Sign Up">
+            <font>
+                <Font size="16.0" />
+            </font>
+        </JFXButton>
+        <JFXButton layoutX="486.0" layoutY="14.0" prefHeight="40.0" prefWidth="100.0" style="-fx-background-color: #03dac6;" text="Login">
+            <font>
+                <Font size="16.0" />
+            </font>
+        </JFXButton>
+      <JFXPasswordField focusColor="#03dac6" layoutX="153.0" layoutY="240.0" prefWidth="300.0" promptText="Re-enter Password" style="-fx-text-fill: #ffffff;">
+         <font>
+            <Font size="16.0" />
+         </font>
+      </JFXPasswordField>
+      <ImageView fitHeight="42.0" fitWidth="50.0" layoutX="108.0" layoutY="238.0" pickOnBounds="true" preserveRatio="true">
+         <image>
+            <Image url="@../assets/lock.png" />
+         </image>
+      </ImageView>
+    </children>
+</AnchorPane>
